requirements/requirements.txt
Django==4.2.1
django-filter==23.2
Pillow==9.5.0
requests==2.30.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
django-environ==0.10.0
django-allauth==0.54.0
psycopg2==2.9.6
dry-rest-permissions==0.1.10
beautifulsoup4==4.12.2
ebaysdk==2.2.0
selenium==4.10.0
html5lib==1.1
